tg_bot_notifications_api/requirements.txt
python-dotenv==0.21.1
aiogram==3.18.0
aiohttp==3.11.12